Refactors song data structure
Replaces the "tom" and "cantor" fields with a new "tons" array field. This allows for multiple singers and their respective keys to be associated with a single song.

Updates the filtering logic to accommodate the new data structure.  The song list view is also updated to display the new "tons" information correctly.